refactor(core): 重构核心模块并移除冗余依赖
- 移除了多个不再使用的 crates 包引用，包括 adapter_api、bot、builder 等
- 更新了 release-please 配置以反映新的包结构
- 合并了一些功能模块，例如将 event_bus 和 matcher 等整合进 bus 模块
- 清理了 Cargo.lock 中过时的依赖项条目
- 在控制台适配器中修复了克隆变量作用域问题，并标准化消息创建宏调用参数
- 调整了工作区成员配置，去除了 dep 目录
- 引入了 tokio-macros 并更新了 puniyu_types 的依赖关系

+++ crates/puniyu_macros/src/lib.rs
@@ -530,7 +530,7 @@ pub fn command(args: TokenStream, item: TokenStream) -> TokenStream {
 				vec![#(#command_args.to_string()),*]
 			}
 
-			fn rank(&self) -> usize {
+			fn rank(&self) -> u64 {
 				#command_rank.to_string().parse().unwrap_or(100)
 			}
